Update app/package name; Add app icon; Add Transaction api; Add StatAdapter; Add header interceptor; Update Profile;

+++ app/src/main/java/domain/model/Country.kt
@@ -1,7 +1,7 @@
 package domain.model
 
 data class Country(
-    val id:Int,
-    val name:String,
-    val image:String
+    val id: Long,
+    val name: String,
+    val image: String
 )
